[Problem 1] brute_force: Fix comparison of integers with different signs

+++ src/1/brute_force.c
@@ -4,8 +4,8 @@ int*
 twoSum(int* nums, int numsSize, int target, int* returnSize)
 {
 
-    for (size_t i = 0; i < numsSize; i++) {
-        for (size_t j = i + 1; j < numsSize; j++) {
+    for (int i = 0; i < numsSize; i++) {
+        for (int j = i + 1; j < numsSize; j++) {
             if (nums[i] + nums[j] == target) {
                 *returnSize = 2;
                 int* res = calloc(2, sizeof(int));
